+++ src/test/java/com/programmers/voucher/global/GlobalApiControllerAdviceTest.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.programmers.voucher.domain.customer.controller.CustomerApiController;
 import com.programmers.voucher.domain.customer.dto.request.CustomerCreateRequest;
 import com.programmers.voucher.domain.customer.service.CustomerService;
+import com.programmers.voucher.global.advice.GlobalApiControllerAdvice;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
